backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
red is==5.0.1
httpx==0.25.2
openai==1.3.0
langchain==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
aioredis==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
PyJWT==2.8.1
email-validator==2.1.0
